# cmd/api/docs/swagger.yaml
definitions:
...
  domain.Brand:
    properties:
      brand_Id:
        type: integer
      brand_description:
        type: string
      brand_name:
        type: string
      discount:
        type: integer
    type: object
...
  domain.Model:
    properties:
      model_id:
        type: integer
      model_name:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server  server. You can visit the GitHub repository
    at https://github.com/prasanth-pn/clean-code-architecture-ecommerce
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin ecommerce API
  version: "1.0"
paths:
  /admin/add/brands:
    post:
      operationId: AddBrand by admin
      parameters:
      - description: Add brand
        in: body
        name: Brand
        required: true
        schema:
          $ref: '#/definitions/domain.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add Brand for admin
      tags:
      - Brand
  /admin/add/category:
    post:
      operationId: AddCategory by admin
      parameters:
      - description: select image
        in: formData
        name: image
        required: true
        type: file
      - description: AdminAddCategory
        in: formData
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List user for admin
      tags:
      - Admin
  /admin/add/models:
    post:
      operationId: AddModel by admin
      parameters:
      - description: Add Model
        in: body
        name: Brand
        required: true
        schema:
          $ref: '#/definitions/domain.Model'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Addmodel for admin
      tags:
      - Model
...
  /admin/list-blockedusers:
    get:
      operationId: ListBlockUser by admin
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: pagesize
        in: query
        name: pagesize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: ListBlockUsers for admin
      tags:
      - USERMANAGEMENT
  /admin/list-orders:
    get:
      operationId: ListOrders by admin
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: pagesize
        in: query
        name: pagesize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: ListOrders for admin
      tags:
      - OrderManagement
...
  /admin/list/users:
    get:
      operationId: ListUsers by admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List user for admin
      tags:
      - Admin
...
  /admin/search-user/name:
    get:
      operationId: SearchUserByName by admin
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: pagesize
        in: query
        name: pagesize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: SearchUserByName for admin
      tags:
      - USERMANAGEMENT
...
  /admin/user/block:
    patch:
      operationId: BlockUser by admin
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Blockuser for admin
      tags:
      - USERMANAGEMENT
  /admin/user/unblock:
    patch:
      operationId: UblockUser by admin
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Unblock for admin
      tags:
      - USERMANAGEMENT
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"